Moved Segregated Profiles tests to a dedicated XCtest file
